requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
python-multipart
faiss-cpu
numpy
sentence-transformers
PyMuPDF
slowapi
json-repair